flight-api-main/requirements.txt
api-analytics==1.1.4
fastapi==0.100.0
python-dotenv==1.0.0
requests==2.31.0
slack-notifications==0.2.2
slowapi==0.1.8
urllib3==2.0.3
uvicorn==0.22.0
gunicorn==21.2.0
jsonpickle==3.0.2
